third_party/blink/web_tests/platform/generic/fast/js/pic/cached-deleted-properties-expected.txt

This test verifies that property caching behaves correctly in the face of deleted properties. If the test passes, you'll see a series of PASS messages below.

PASS: Array.equal(properties, expectedProperties) should be true and is.
PASS: Array.equal(properties, expectedProperties) should be true and is.
PASS: Array.equal(properties, expectedProperties) should be true and is.
PASS: o.x should be 1 and is.
PASS: o.x should be 1 and is.
PASS: o.x should be 1 and is.
